LPS-96040 Better name
