--- modules/apps/portal-vulcan/portal-vulcan-impl/src/test/java/com/liferay/portal/vulcan/internal/util/servlet/MockHttpServletRequest.java
@@ -1,391 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.vulcan.internal.util.servlet;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.UnsupportedEncodingException;
-
-import java.security.Principal;
-
-import java.util.Collection;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
-
-import javax.servlet.AsyncContext;
-import javax.servlet.DispatcherType;
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletInputStream;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.Cookie;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-import javax.servlet.http.Part;
-
-/**
- * @author Ivica Cardic
- */
-public class MockHttpServletRequest implements HttpServletRequest {
-
-	public MockHttpServletRequest() {
-	}
-
-	public MockHttpServletRequest(String fieldName, String... parameters) {
-		for (int i = 0; i < (parameters.length - 1); i++) {
-			_parameters.put(fieldName + "." + parameters[i], parameters[i + 1]);
-		}
-	}
-
-	@Override
-	public boolean authenticate(HttpServletResponse httpServletResponse)
-		throws IOException, ServletException {
-
-		return false;
-	}
-
-	@Override
-	public AsyncContext getAsyncContext() {
-		return null;
-	}
-
-	@Override
-	public Object getAttribute(String name) {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getAttributeNames() {
-		return null;
-	}
-
-	@Override
-	public String getAuthType() {
-		return null;
-	}
-
-	@Override
-	public String getCharacterEncoding() {
-		return null;
-	}
-
-	@Override
-	public int getContentLength() {
-		return 0;
-	}
-
-	@Override
-	public String getContentType() {
-		return null;
-	}
-
-	@Override
-	public String getContextPath() {
-		return null;
-	}
-
-	@Override
-	public Cookie[] getCookies() {
-		return new Cookie[0];
-	}
-
-	@Override
-	public long getDateHeader(String name) {
-		return 0;
-	}
-
-	@Override
-	public DispatcherType getDispatcherType() {
-		return null;
-	}
-
-	@Override
-	public String getHeader(String name) {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getHeaderNames() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getHeaders(String name) {
-		return null;
-	}
-
-	@Override
-	public ServletInputStream getInputStream() throws IOException {
-		return null;
-	}
-
-	@Override
-	public int getIntHeader(String name) {
-		return 0;
-	}
-
-	@Override
-	public String getLocalAddr() {
-		return null;
-	}
-
-	@Override
-	public Locale getLocale() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<Locale> getLocales() {
-		return null;
-	}
-
-	@Override
-	public String getLocalName() {
-		return null;
-	}
-
-	@Override
-	public int getLocalPort() {
-		return 0;
-	}
-
-	@Override
-	public String getMethod() {
-		return null;
-	}
-
-	@Override
-	public String getParameter(String name) {
-		return _parameters.get(name);
-	}
-
-	@Override
-	public Map<String, String[]> getParameterMap() {
-		return null;
-	}
-
-	@Override
-	public Enumeration<String> getParameterNames() {
-		return null;
-	}
-
-	@Override
-	public String[] getParameterValues(String name) {
-		return new String[0];
-	}
-
-	@Override
-	public Part getPart(String name) throws IOException, ServletException {
-		return null;
-	}
-
-	@Override
-	public Collection<Part> getParts() throws IOException, ServletException {
-		return null;
-	}
-
-	@Override
-	public String getPathInfo() {
-		return null;
-	}
-
-	@Override
-	public String getPathTranslated() {
-		return null;
-	}
-
-	@Override
-	public String getProtocol() {
-		return null;
-	}
-
-	@Override
-	public String getQueryString() {
-		return null;
-	}
-
-	@Override
-	public BufferedReader getReader() throws IOException {
-		return null;
-	}
-
-	@Override
-	public String getRealPath(String path) {
-		return null;
-	}
-
-	@Override
-	public String getRemoteAddr() {
-		return null;
-	}
-
-	@Override
-	public String getRemoteHost() {
-		return null;
-	}
-
-	@Override
-	public int getRemotePort() {
-		return 0;
-	}
-
-	@Override
-	public String getRemoteUser() {
-		return null;
-	}
-
-	@Override
-	public RequestDispatcher getRequestDispatcher(String path) {
-		return null;
-	}
-
-	@Override
-	public String getRequestedSessionId() {
-		return null;
-	}
-
-	@Override
-	public String getRequestURI() {
-		return null;
-	}
-
-	@Override
-	public StringBuffer getRequestURL() {
-		return null;
-	}
-
-	@Override
-	public String getScheme() {
-		return null;
-	}
-
-	@Override
-	public String getServerName() {
-		return null;
-	}
-
-	@Override
-	public int getServerPort() {
-		return 0;
-	}
-
-	@Override
-	public ServletContext getServletContext() {
-		return null;
-	}
-
-	@Override
-	public String getServletPath() {
-		return null;
-	}
-
-	@Override
-	public HttpSession getSession() {
-		return null;
-	}
-
-	@Override
-	public HttpSession getSession(boolean create) {
-		return null;
-	}
-
-	@Override
-	public Principal getUserPrincipal() {
-		return null;
-	}
-
-	@Override
-	public boolean isAsyncStarted() {
-		return false;
-	}
-
-	@Override
-	public boolean isAsyncSupported() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromCookie() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromUrl() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdFromURL() {
-		return false;
-	}
-
-	@Override
-	public boolean isRequestedSessionIdValid() {
-		return false;
-	}
-
-	@Override
-	public boolean isSecure() {
-		return false;
-	}
-
-	@Override
-	public boolean isUserInRole(String role) {
-		return false;
-	}
-
-	@Override
-	public void login(String username, String password)
-		throws ServletException {
-	}
-
-	@Override
-	public void logout() throws ServletException {
-	}
-
-	@Override
-	public void removeAttribute(String name) {
-	}
-
-	@Override
-	public void setAttribute(String name, Object value) {
-	}
-
-	@Override
-	public void setCharacterEncoding(String characterEncoding)
-		throws UnsupportedEncodingException {
-	}
-
-	@Override
-	public AsyncContext startAsync() throws IllegalStateException {
-		return null;
-	}
-
-	@Override
-	public AsyncContext startAsync(
-			ServletRequest servletRequest, ServletResponse servletResponse)
-		throws IllegalStateException {
-
-		return null;
-	}
-
-	private Map<String, String> _parameters = new HashMap<>();
-
-}
